Created compile-larceny.scm
Due to the difficulties I've been having using (command-line) in Windows,
I've created compile-larceny to read compilation options from a file and then
call compile-r7rs based on those. The idea is to use a batch file to write those
options to a file, then call upon compile-larceny to actually compile it, then
make the batch file clean up after itself.

Right now, it doesn't support -D stated in the SRFI, but it does accept the
-I, -A, and -o compilation switches